requirements.txt
discord.py
requests
python-qbittorrent